+++ user-licence-lifecycle-manager/src/main/java/de/intension/authentication/authenticators/licence/LicenceConnectAuthenticator.java
@@ -5,8 +5,8 @@ import de.intension.authentication.authenticators.jpa.LicenceJpaProvider;
 import de.intension.authentication.authenticators.jpa.entity.LicenceEntity;
 import de.intension.protocol.oidc.mappers.HmacPairwiseSubMapper;
 import de.intension.protocol.oidc.mappers.HmacPairwiseSubMapperHelper;
-import de.intension.rest.LicenceConnectRestClient;
-import de.intension.rest.model.LicenceRequest;
+import de.intension.rest.licence.client.LicenceConnectRestClient;
+import de.intension.rest.licence.model.LicenceRequest;
 import jakarta.ws.rs.core.Response;
 import lombok.Getter;
 import org.jboss.logging.Logger;
