+++ user-licence-lifecycle-manager/src/main/java/de/intension/listener/RemoveUserOnLogOutEventListenerProvider.java
@@ -4,8 +4,8 @@ import de.intension.authentication.authenticators.jpa.LicenceJpaProvider;
 import de.intension.protocol.oidc.mappers.HmacPairwiseSubMapper;
 import de.intension.protocol.oidc.mappers.HmacPairwiseSubMapperHelper;
 import de.intension.resources.admin.DeletableUserType;
-import de.intension.rest.LicenceConnectRestClient;
-import de.intension.rest.model.RemoveLicenceRequest;
+import de.intension.rest.licence.client.LicenceConnectRestClient;
+import de.intension.rest.licence.model.RemoveLicenceRequest;
 import de.intension.spi.RestClientProvider;
 import jakarta.persistence.EntityManager;
 import lombok.Getter;
