+++ user-licence-lifecycle-manager/src/main/java/de/intension/spi/RestClientProvider.java
@@ -1,6 +1,6 @@
 package de.intension.spi;
 
-import de.intension.rest.LicenceConnectRestClient;
+import de.intension.rest.licence.client.LicenceConnectRestClient;
 import org.keycloak.provider.Provider;
 
 public interface RestClientProvider
