+++ user-licence-lifecycle-manager/src/main/resources/META-INF/services/de.intension.spi.RestClientProviderFactory
@@ -1,1 +1,1 @@
-de.intension.rest.DefaultRestClientProviderFactory
+de.intension.rest.licence.client.DefaultRestClientProviderFactory
