+++ user-licence-lifecycle-manager/src/main/resources/META-INF/services/org.keycloak.services.resource.RealmResourceProviderFactory
@@ -1,1 +1,1 @@
-de.intension.rest.LicenceResourceProviderFactory
+de.intension.rest.licence.LicenceResourceProviderFactory
